refactor - changed structure of code
