+++ recommender/static/summoners.txt
