+++ recommender/static/swagger.json
